// src/lib/enums/product.enum.ts
export enum ProductSize {
  SMALL = "SMALL",
  MEDIUM = "MEDIUM",
  LARGE = "LARGE",
  SET = "SET",
}

export enum ProductColor {
  RED = "RED",
  BLUE = "BLUE",
  GREEN = "GREEN",
  YELLOW = "YELLOW",
  BLACK = "BLACK",
  WHITE = "WHITE",
  GREY = "GREY",
  ORANGE = "ORANGE",
  PINK = "PINK",
  BROWN = "BROWN",
  GOLD = "GOLD",
  SILVER = "SILVER",
}

export enum ProductVolume {
  SMALL_ONE = "10-12mm",
  MEDIUM_TWO = "12-14mm",
  LARGE_THREE = "14-16mm",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  SAMSUNG_WATCH = "SAMSUNG_WATCH",
  APPLE_WATCH = "APPLE_WATCH",
  HUAWEI_WATCH = "HUAWEI_WATCH",
  EARPODS = "EARPODS",
  EARBUDS = "EARBUDS",
  OTHERS = "OTHERS",
}